Updated rasa model training to only include nlu training.
Minimal set of dialog management was initially being trained on the rasa model, even though it wasn't being used. Since it is only being used for NLU and not any dialog management, it has been updated to be trained only for the NLU portion. This has reduced the ram usage while the model is loaded by Mykros by approximately 10%.